refactor(swc/plugin_runner): extract common memory utility fn
