+++ crates/swc_plugin_runner/src/imported_fn/comments.rs
@@ -6,10 +6,10 @@ use swc_common::{
     plugin::Serialized,
     BytePos,
 };
-use swc_plugin_proxy::{AllocatedBytesPtr, COMMENTS};
+use swc_plugin_proxy::COMMENTS;
 use wasmer::{LazyInit, Memory, NativeFunc};
 
-use crate::memory_interop::{copy_bytes_into_host, write_into_memory_view};
+use crate::memory_interop::{allocate_return_values_into_guest, copy_bytes_into_host};
 
 /// External environment state for imported (declared in host, injected into
 /// guest) fn for comments proxy.
@@ -97,44 +97,6 @@ where
         }
     }
     default
-}
-
-/// Set `return` value to pass into guest from functions returning values with
-/// non-deterministic size like `Vec<Comment>`. Guest pre-allocates a struct to
-/// contain ptr to the value, host in here allocates guest memory for the actual
-/// value then returns its ptr with length to the preallocated struct.
-fn allocate_return_values_into_guest(
-    memory: &Memory,
-    alloc_guest_memory: &NativeFunc<u32, i32>,
-    allocated_ret_ptr: i32,
-    serialized_bytes: &Serialized,
-) {
-    let serialized_bytes_len = serialized_bytes.as_ref().len();
-
-    let (allocated_ptr, allocated_ptr_len) =
-        write_into_memory_view(memory, serialized_bytes, |_| {
-            // In most cases our host-plugin trampoline works in a way that
-            // plugin pre-allocates
-            // memory before calling host imported fn. But in case of
-            // comments return value is Vec<Comments> which
-            // guest cannot predetermine size to allocate, instead
-            // let host allocate by calling guest's alloc via attached
-            // hostenvironment.
-            alloc_guest_memory
-                .call(
-                    serialized_bytes_len
-                        .try_into()
-                        .expect("Should be able to convert size"),
-                )
-                .expect("Should able to allocate memory in the plugin")
-        });
-
-    // Retuning (allocated_ptr, len) into caller (plugin)
-    let comment_ptr_serialized =
-        Serialized::serialize(&AllocatedBytesPtr(allocated_ptr, allocated_ptr_len))
-            .expect("Should be serializable");
-
-    write_into_memory_view(memory, &comment_ptr_serialized, |_| allocated_ret_ptr);
 }
 
 /// Common logics for add_*_comment/comments.
